- Don't copy the backup file in "LaunchBackupCfg" of "MainWindow.xaml" class, if trying to save the file alredy saved in "appData" folder.

+++ MainWindow.xaml.cs
@@ -502,9 +502,13 @@ namespace PoEWizard
                     };
                     if (sfd.ShowDialog() == true)
                     {
-                        File.Copy(zipPath, sfd.FileName, true);
+                        if (sfd.FileName != zipPath)
+                        {
+                            File.Copy(zipPath, sfd.FileName, true);
+                            File.Delete(zipPath);
+                        }
                     }
-                    File.Delete(zipPath);
+                    else File.Delete(zipPath);
                 }
                 else
                 {
